% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{alter_covariate}
\alias{alter_covariate}
\title{alter_covariate}
\usage{
alter_covariate(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  loop.contribution = T,
  getTargetContribution = F
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{loop.contribution}{}

\item{getTargetContribution}{}
}
\description{
alter_covariate
}
\keyword{internal}
