% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{as.nodeVariable}
\alias{as.nodeVariable}
\title{Assigns an attribute to nodes}
\usage{
as.nodeVariable(values)
}
\arguments{
\item{values}{An attribute in vector form to assign to nodes in a mobility network}
}
\value{
a nodevar object
}
\description{
lalalala this is a test
}
