% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2f_effectFunctions_endogenousCovariateBased.R
\name{associativity_one_GW_covar_bin}
\alias{associativity_one_GW_covar_bin}
\title{associativity_one_GW_covar_bin}
\usage{
associativity_one_GW_covar_bin(
  dep.var = 1,
  resource.attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{resource.attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Do individuals with the same attribute tend to use the same paths and
individuals with different attributes to move to different places?
This statistic gives a positive contribution to all cliques of individuals
with the value 1 of a binary covariate who use the same path.
It's a geometrically-weighted statistic, so the more cliques we add, the
less they matter.
}
\keyword{internal}
