% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2f_effectFunctions_endogenousCovariateBased.R
\name{avoiding_dissimilar_covar_bin}
\alias{avoiding_dissimilar_covar_bin}
\title{avoiding_dissimilar_covar_bin}
\usage{
avoiding_dissimilar_covar_bin(
  dep.var = 1,
  resource.attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Do individuals with different attributes to move to different places?
This statistic gives a negative contribution to pairs of dissimilar
individuals following the same path.
}
\keyword{internal}
