% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2f_effectFunctions_endogenousCovariateBased.R
\name{avoiding_dissimilar_covar_cont}
\alias{avoiding_dissimilar_covar_cont}
\title{avoiding_dissimilar_covar_cont}
\usage{
avoiding_dissimilar_covar_cont(
  dep.var = 1,
  resource.attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{resource.attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Do individuals with different attributes tend to move to different places?
This statistic gives a contribution to all pairs of individuals who use the
same path that is weighted by the absolute difference between their continuous attribute.
}
\keyword{internal}
