% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{binary_transitivity}
\alias{binary_transitivity}
\title{binary_transitivity}
\usage{
binary_transitivity(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  k = NULL,
  l = NULL,
  getTargetContribution = F
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{k}{}

\item{l}{}

\item{getTargetContribution}{}
}
\description{
binary_transitivity
}
\keyword{internal}
