% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c_effectFunctions_concentration.R
\name{concentration_exponent}
\alias{concentration_exponent}
\title{concentration_exponent}
\usage{
concentration_exponent(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  exponent = 0.5
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{exponent}{}
}
\description{
concentration_exponent
}
\keyword{internal}
