% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c_effectFunctions_concentration.R
\name{concentration_GW}
\alias{concentration_GW}
\title{concentration_GW}
\usage{
concentration_GW(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  lambda = 2
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{lambda}{}
}
\description{
concentration_GW
}
\keyword{internal}
