% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c_effectFunctions_concentration.R
\name{concentration_GW_dyad_covar_bin}
\alias{concentration_GW_dyad_covar_bin}
\title{concentration_GW_dyad_covar_bin}
\usage{
concentration_GW_dyad_covar_bin(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE,
  alpha = 2
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{alpha}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Are bandwagon effects (concentration) particularly prevalent between locations
that share characteristics as encoded in a binary dyadic covariate? E.g., do
workers follow the moves of other workers mainly in case they go to organisations
in the same region?
}
\keyword{internal}
