% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c_effectFunctions_concentration.R
\name{concentration_norm_squared}
\alias{concentration_norm_squared}
\title{concentration_norm_squared}
\usage{
concentration_norm_squared(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Is there a bandwagon effect in mobility, i.e. do mobile individuals move to
locations that are the destination of many others from their origin?
}
\keyword{internal}
