% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c_effectFunctions_concentration.R
\name{concentration_rankGW}
\alias{concentration_rankGW}
\title{concentration_rankGW}
\usage{
concentration_rankGW(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE,
  lambda = 2
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{lambda}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Is there a bandwagon effect in mobility, i.e. do mobile individuals move to locations
that are the destination of many others from their origin? The functional form of this
statistic assumes that there are increasing returns to choosing more populated paths
but the return of using a path only depends on the ranking of this path in terms of number
of people using it among all paths (and not the actual numbers of people on this path).
}
\keyword{internal}
