% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c_effectFunctions_concentration.R
\name{concentration_sigmoid_dyad_covar}
\alias{concentration_sigmoid_dyad_covar}
\title{concentration_sigmoid_dyad_covar}
\usage{
concentration_sigmoid_dyad_covar(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  alpha
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{alpha}{}
}
\description{
concentration_sigmoid_dyad_covar
}
\keyword{internal}
