% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{createAlgorithm}
\alias{createAlgorithm}
\title{createAlgorithm}
\usage{
createAlgorithm(
  state,
  effects,
  burnInN1 = NULL,
  iterationsN1 = NULL,
  thinningN1 = NULL,
  gainN1 = 0.1,
  burnInN2 = NULL,
  nsubN2 = 4,
  initGain = 0.2,
  thinningN2 = NULL,
  initialIterationsN2 = 25,
  iterationsN3 = 500,
  burnInN3 = NULL,
  thinningN3 = NULL,
  allowLoops = TRUE
)
}
\arguments{
\item{allowLoops}{}
}
\description{
Specifies the algorithm used in later model estimation based on characteristics
of the state and the effects.
}
\examples{
# define algorithm based on state and effects characteristics
myAlg <- createAlgorithm(myState, myEffects)
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}, \code{\link[=createEffectsObject]{createEffectsObject()}}, \code{\link[=estimateMobilityNetwork]{estimateMobilityNetwork()}}
}
