% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_wrapperFunctions.R
\name{createEffects}
\alias{createEffects}
\title{createEffects}
\usage{
createEffects(state)
}
\arguments{
\item{state}{The state to which the model applies.}
}
\value{
An empty effects object of class effectsList.monan
}
\description{
Generates an empty effects object to which new effects can be added
consecutively
}
\examples{
#' myE1 <- createEffects(myState)
}
