% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{createEffectsObject}
\alias{createEffectsObject}
\title{Title}
\usage{
createEffectsObject(effectInit, checkProcessState = NULL)
}
\arguments{
\item{checkProcessState}{}
}
\description{
Title
}
\examples{
# create an effects object
exampleEffects <- createEffectsObject(
  list(
    list("loops"),
    list("min_reciprocity"),
    list("loops_resource_covar", resource.attribute.index = "resVarCat")
  )
)
}
