% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{createNodeVariable}
\alias{createNodeVariable}
\title{Title}
\usage{
createNodeVariable(
  values,
  range = NULL,
  nodeSet = "actors",
  addSame = F,
  addSim = F
)
}
\arguments{
\item{addSim}{}
}
\description{
Title
}
\examples{
# create an object of class nodeVar.monan
nodeVarCat_NV <- createNodeVariable(nodeVarCat, nodeSet = "location")
nodeVarCont_NV <- createNodeVariable(nodeVarCont, nodeSet = "location", addSim = TRUE)
resVarCat_NV <- createNodeVariable(resVarCat, nodeSet = "people")
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}
}
