% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{dyadic_covariate_tie_weights_sigmoid}
\alias{dyadic_covariate_tie_weights_sigmoid}
\title{dyadic_covariate_tie_weights_sigmoid}
\usage{
dyadic_covariate_tie_weights_sigmoid(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  alpha
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{alpha}{}
}
\description{
dyadic_covariate_tie_weights_sigmoid
}
\keyword{internal}
