% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R, R/4_auxiliaryFunctions.R
\name{estimateMobilityNetwork}
\alias{estimateMobilityNetwork}
\alias{estimateDistributionNetwork}
\alias{print.result.monan}
\title{estimateMobilityNetwork}
\usage{
estimateMobilityNetwork(
  dep.var,
  state,
  cache,
  effects,
  initialParameters = NULL,
  prevAns = NULL,
  burnInN1 = 500,
  iterationsN1 = 200,
  thinningN1 = 20,
  gainN1 = 1,
  burnInN2 = burnInN1,
  nsubN2 = 6,
  initGain = 0.02,
  thinningN2 = 100,
  initialIterationsN2 = 50,
  iterationsN3 = 1000,
  burnInN3 = 2000,
  thinningN3 = 30,
  allowLoops = NULL,
  parallel = FALSE,
  cpus = 3,
  verbose = FALSE,
  returnDeps = FALSE,
  multinomialProposal = FALSE,
  fish = FALSE
)

estimateDistributionNetwork(
  dep.var,
  state,
  cache,
  effects,
  initialParameters = NULL,
  prevAns = NULL,
  burnInN1 = 500,
  iterationsN1 = 200,
  thinningN1 = 20,
  gainN1 = 1,
  burnInN2 = burnInN1,
  nsubN2 = 6,
  initGain = 0.02,
  thinningN2 = 100,
  initialIterationsN2 = 50,
  iterationsN3 = 1000,
  burnInN3 = 2000,
  thinningN3 = 30,
  allowLoops = NULL,
  parallel = FALSE,
  cpus = 3,
  verbose = FALSE,
  returnDeps = FALSE,
  multinomialProposal = FALSE,
  fish = FALSE
)

\method{print}{result.monan}(x, covMat = FALSE, ...)
}
\arguments{
\item{fish}{}

\item{x}{An ans object resulting from an estimation with estimateMobilityNetwork()}

\item{covMat}{Logical indicating whether the covariance matrix should be printed}

\item{...}{For internal use only}
}
\description{
estimateMobilityNetwork

estimateDistributionNetwork

print.result.monan
}
\examples{
\dontrun{
# estimate mobility network
myResDN <- estimateMobilityNetwork(myDependentVariable,
                                   myState, myCache, myEffects,
                                   initialParameters = NULL,
                                   burnInN1 = 1500, iterationsN1 = 50, 
                                   thinningN1 = 750, gainN1 = 0.1,
                                   burnInN2 = 7500, nsubN2 = 4, 
                                   initGain = 0.2, thinningN2 = 1500,
                                   initialIterationsN2 = 25,
                                   iterationsN3 = 500, burnInN3 = 7500, 
                                   thinningN3 = 3750,
                                   parallel = TRUE, cpus = 4,
                                   allowLoops = TRUE,
                                   verbose = TRUE,
                                   returnDeps = TRUE,
                                   multinomialProposal = FALSE,
                                   fish = FALSE
)

myResDN

# estimate mobility network again based on previous results to improve convergence
myResDN <- estimateMobilityNetwork(myDependentVariable,
                                   myState, myCache, myEffects,
                                   prevAns = myResDN,
                                   burnInN1 = 1500, iterationsN1 = 50, 
                                   thinningN1 = 750, gainN1 = 0.1,
                                   burnInN2 = 7500, nsubN2 = 4, 
                                   initGain = 0.2, thinningN2 = 3000,
                                   initialIterationsN2 = 40,
                                   iterationsN3 = 500, burnInN3 = 15000, 
                                   thinningN3 = 7500,
                                   parallel = TRUE, cpus = 4,
                                   allowLoops = TRUE,
                                   verbose = TRUE,
                                   returnDeps = TRUE,
                                   multinomialProposal = FALSE,
                                   fish = FALSE
)
}
myResDN
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}, \code{\link[=createWeightedCache]{createWeightedCache()}},
\code{\link[=createEffectsObject]{createEffectsObject()}}
}
