% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\alias{mobilityEdgelist}
\alias{nodeVarCat}
\alias{nodeVarCont}
\alias{resVarCat}
\title{Example Data for the MoNAn Package}
\format{
\code{mobilityEdgelist}
A data frame with 743 rows and 2 columns.

\code{nodeVarCat}
An object with 17 values.

\code{nodeVarCont}
An object with 17 values.

\code{resVarCat}
An object with 742 values.
}
\usage{
mobilityEdgelist

nodeVarCat

nodeVarCont

resVarCat
}
\description{
These are example data for the MoNAn package and can be used to estimate a
mobility network. The following objects are provided for this purpose:

\describe{
\item{mobilityEdgelist}{blabla}
\item{nodeVarCat}{blabla}
\item{nodeVarCont}{blabla}
\item{resVarCat}{blabla}
}
}
\keyword{datasets}
