% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleProducts.R
\docType{data}
\name{exampleProducts}
\alias{exampleProducts}
\alias{exampleState}
\alias{exampleDependentVariable}
\alias{exampleCache}
\alias{exampleEffects}
\alias{exampleResDN}
\title{Example Products for the MoNAn Package}
\format{
\code{exampleState}
An object of class processState.monan.

\code{exampleDependentVariable}
An object of class character.

\code{exampleCache}
An object of class list.

\code{exampleEffects}
An object of class "effectsList.monan".

\code{exampleResDN}
An object of class "result.monan".
}
\usage{
exampleState

exampleDependentVariable

exampleCache

exampleEffects

exampleResDN
}
\description{
These are example products for the MoNAn package and can be used in order not
to run all precedent functions and thus save time. The following products are
provided:

\describe{
\item{exampleState}{blabla}
\item{exampleDependentVariable}{blabla}
\item{exampleCache}{blabla}
\item{exampleEffects}{blabla}
\item{exampleResDN}{blabla}
}
}
\keyword{datasets}
