% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{extractTraces}
\alias{extractTraces}
\title{Title}
\usage{
extractTraces(dep.var, ans, effects)
}
\arguments{
\item{effects}{}
}
\description{
Title
}
\examples{
# regression diagnostics
traces <- extractTraces(dependentVariable, resDN, exampleEffects)
plot(traces)
}
\seealso{
\code{\link{estimateMobilityNetwork}}, \code{\link{createEffectsObject}}
}
