% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_gofParamFunctions.R
\name{getIndegree}
\alias{getIndegree}
\title{getIndegree}
\usage{
getIndegree(cache, dep.var, lvls, ...)
}
\arguments{
\item{cache}{Current Cache.}

\item{dep.var}{Dependent Variable.}

\item{lvls}{Levels for which the function calculates values.}

\item{...}{Additional parameters.}
}
\value{
The degree distribution of simulated networks.
}
\description{
Calculates the weighted indegree distribution of all locations in the
network.
The weighted indegree is simply the column sum of the mobility table.
}
\seealso{
\code{\link[=gofMobilityNetwork]{gofMobilityNetwork()}}
}
\keyword{internal}
