% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_auxiliaryFunctions.R
\name{getMultinomialStatistics}
\alias{getMultinomialStatistics}
\title{Title}
\usage{
getMultinomialStatistics(state, cache, effects, dep.var)
}
\arguments{
\item{state}{}

\item{cache}{}

\item{effects}{}

\item{dep.var}{}
}
\description{
Title
}
\examples{
myStatisticsFrame <- getMultinomialStatistics(myState, myCache, myEffects, myDependentVariable)
}
