% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_gofParamFunctions.R
\name{getTieWeights}
\alias{getTieWeights}
\title{getTieWeights}
\usage{
getTieWeights(cache, dep.var, lvls, ...)
}
\arguments{
\item{cache}{Current Cache.}

\item{dep.var}{Dependent Variable.}

\item{lvls}{Levels for which the function calculates values.}

\item{...}{Additional parameters.}
}
\value{
None.
}
\description{
getTieWeights
}
\seealso{
\code{\link[=gofMobilityNetwork]{gofMobilityNetwork()}}
}
\keyword{internal}
