% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{gofDistributionNetwork}
\alias{gofDistributionNetwork}
\title{Title}
\usage{
gofDistributionNetwork(
  ans,
  simulations,
  gofFunction,
  dep.var = NULL,
  lvls = NULL
)
}
\arguments{
\item{lvls}{}
}
\description{
Title
}
