% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{in_proportion_exponent_covar_bin}
\alias{in_proportion_exponent_covar_bin}
\title{in_proportion_exponent_covar_bin}
\usage{
in_proportion_exponent_covar_bin(
  dep.var = 1,
  resource.attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  exponent = 2
)
}
\arguments{
\item{dep.var}{}

\item{resource.attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{exponent}{}
}
\description{
in_proportion_exponent_covar_bin
}
\keyword{internal}
