% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{in_weights_AC}
\alias{in_weights_AC}
\title{in_weights_AC}
\usage{
in_weights_AC(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE,
  alpha = 2
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{alpha}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Is there a preferential attachment in the mobility network, i.e., do individuals
move particularly to popular destinations?
The geometrically weighted version avoids degeneracy.
}
\keyword{internal}
