% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{in_weights_exponent}
\alias{in_weights_exponent}
\title{in_weights_exponent}
\usage{
in_weights_exponent(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  exponent = 0.5
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{exponent}{}
}
\description{
in_weights_exponent
}
\keyword{internal}
