% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2f_effectFunctions_endogenousCovariateBased.R
\name{joining_similar_avoiding_dissimilar_covar_bin}
\alias{joining_similar_avoiding_dissimilar_covar_bin}
\title{joining_similar_avoiding_dissimilar_covar_bin}
\usage{
joining_similar_avoiding_dissimilar_covar_bin(
  dep.var = 1,
  resource.attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{resource.attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Do individuals with the same attribute tend to use the same paths and
individuals with different attributes to move to different places?
This statistic gives a positive contribution to all pairs of individuals
with the same (binary) covariate who use the same path and a negative one
to pairs of dissimilar individuals following the same path.
}
\keyword{internal}
