% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2f_effectFunctions_endogenousCovariateBased.R
\name{joining_similar_avoiding_dissimilar_covar_cont}
\alias{joining_similar_avoiding_dissimilar_covar_cont}
\title{joining_similar_avoiding_dissimilar_covar_cont}
\usage{
joining_similar_avoiding_dissimilar_covar_cont(
  dep.var = 1,
  resource.attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{resource.attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Do individuals with the same attribute tend to use the same paths and
individuals with different attributes to move to different places?
This statistic gives a contribution to all pairs of individuals who use the
same path that is weighted by the similarity between their continuous attribute.
This similarity weight is measured as the range of the attribute, minus twice
the absolute difference between their covariates, normalized by the range (the
weight of a pair is equal to 1 if both individuals have the same attribute,
and -1 if the absolute difference between their attributes is equal to the range).
}
\keyword{internal}
