% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_effectFunctions_loops.R
\name{loops_GW}
\alias{loops_GW}
\title{loops_GW}
\usage{
loops_GW(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  alpha = 2
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{alpha}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Do individuals stay in their current location, in case many other from their
current location also stay? This effect tests whether the ‘benefit’ of staying
in the origin location depends on the number of others also staying. Note that
this effect should be modelled alongside the loops effect.
}
\keyword{internal}
