% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_effectFunctions_loops.R
\name{loops_GW_prop}
\alias{loops_GW_prop}
\title{loops_GW_prop}
\usage{
loops_GW_prop(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  half.cont = 0.1
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{half.cont}{}
}
\value{
None.
}
\description{
loops_GW_prop
}
\keyword{internal}
