% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{loops_neg_exp}
\alias{loops_neg_exp}
\title{loops_neg_exp}
\usage{
loops_neg_exp(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  half.cont = 0.1
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{half.cont}{}
}
\description{
loops_neg_exp
}
\keyword{internal}
