% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_effectFunctions_loops.R
\name{loops_node_covar}
\alias{loops_node_covar}
\title{loops_node_covar}
\usage{
loops_node_covar(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
None.
}
\description{
loops_node_covar
}
\keyword{internal}
