% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_effectFunctions_loops.R
\name{loops_x_loops_additional_origin}
\alias{loops_x_loops_additional_origin}
\title{loops_x_loops_additional_origin}
\usage{
loops_x_loops_additional_origin(
  dep.var = 1,
  resource.attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{resource.attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
This effect is specified for cases in which individuals have more than one
origin. The additional origin not specified in the mobility data is included
as a resource.attribute.index.
The question modeled is: Do individuals stay in the additional location of
origin if this is additionally their origin as specified in the mobility data,
compared to going to a different location?
}
\keyword{internal}
