% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{min_reciprocity}
\alias{min_reciprocity}
\title{min_reciprocity}
\usage{
min_reciprocity(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\description{
min_reciprocity
}
\keyword{internal}
