% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilityData.R
\docType{data}
\name{mobilityData}
\alias{mobilityData}
\alias{mobilityEdgelist}
\alias{orgRegion}
\alias{orgSize}
\alias{indSex}
\title{Example Data for the MoNAn Package}
\format{
\code{mobilityEdgelist}
A data frame with 743 rows and 2 columns.

\code{orgRegion}
An object with 17 values.

\code{orgSize}
An object with 17 values.

\code{indSex}
An object with 742 values.
}
\description{
These are example data for the MoNAn package and can be used to estimate a
mobility network. The following objects are provided for this purpose:

\describe{
\item{\code{mobilityEdgelist}}{This data frame represents the origin at time 1
(first column) as well as the destination at time 2 (second column) for
each of the 742 individuals between the 17 organisations. Individuals could
also stay in their organisation meaning that origin equals destination.}
\item{\code{orgRegion}}{This characteristic describes whether the resp.
organisation is located on the northern (1) or southern (0) island.}
\item{\code{orgSize}}{This composite measure represents the size of each
organisation including its number of workers, assets, and revenue.}
\item{\code{indSex}}{This characteristic represents each individual's gender.}
}
}
\keyword{datasets}
