% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoNAn-package.R
\docType{package}
\name{MoNAn-package}
\alias{MoNAn}
\alias{MoNAn-package}
\title{MoNAn: Mobility Network Analysis}
\description{
Implements the method to analyse weighted mobility networks or distribution networks as outlined in: Block, P., Stadtfeld, C., & Robins, G. (2022). A statistical model for the analysis of mobility tables as weighted networks with an application to faculty hiring networks. Social Networks, 68, 264-278.
}
\author{
\strong{Maintainer}: Per Block \email{block@soziologie.uzh.ch}

Authors:
\itemize{
  \item Christoph Stadtfeld
}

}
\keyword{internal}
