% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_auxiliaryFunctions.R
\name{print.effectsList.monan}
\alias{print.effectsList.monan}
\title{print.effectsList.monan}
\usage{
\method{print}{effectsList.monan}(x)
}
\arguments{
\item{x}{An object of class "effectsList.monan".}
}
\value{
The function \code{print.effectsList.monan} gives an overview of the
specified effects.
}
\description{
print.effectsList.monan
}
\examples{
myEffects
}
