% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_auxiliaryFunctions.R
\name{print.processState.monan}
\alias{print.processState.monan}
\title{print.processState.monan}
\usage{
\method{print}{processState.monan}(x)
}
\arguments{
\item{x}{An object of class "processState.monan".}
}
\value{
The function \code{print.processState.monan} gives an overview of the information
included in the state object.
}
\description{
print.processState.monan
}
\examples{
myState
}
