% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_effectFunctions_reciprocity.R
\name{reciprocity_AC_dyad_covar}
\alias{reciprocity_AC_dyad_covar}
\title{reciprocity_AC_dyad_covar}
\usage{
reciprocity_AC_dyad_covar(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE,
  alpha = 2
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{alpha}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Is reciprocity in mobility particularly prevalent between locations weighted
by a dyadic covariate? E.g., do workers move
to organisations similar in some way to ego’s origin?
}
\keyword{internal}
