% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_effectFunctions_reciprocity.R
\name{reciprocity_basic}
\alias{reciprocity_basic}
\title{reciprocity_basic}
\usage{
reciprocity_basic(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
None.
}
\description{
reciprocity_basic
}
\keyword{internal}
