% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_effectFunctions_reciprocity.R
\name{reciprocity_GW}
\alias{reciprocity_GW}
\title{reciprocity_GW}
\usage{
reciprocity_GW(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE,
  lambda = 2
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{lambda}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Do individuals move to destinations that send many individuals to ego’s origin?
The number of incoming individuals has decreasing returns, that is, every
additionally incoming individual influences ego’s choice less.
}
\keyword{internal}
