% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_effectFunctions_reciprocity.R
\name{reciprocity_GW_dyad_covar_bin}
\alias{reciprocity_GW_dyad_covar_bin}
\title{reciprocity_GW_dyad_covar_bin}
\usage{
reciprocity_GW_dyad_covar_bin(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  lambda = 2
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{lambda}{}
}
\value{
None.
}
\description{
reciprocity_GW_dyad_covar_bin
}
\keyword{internal}
