% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_effectFunctions_reciprocity.R
\name{reciprocity_min_dyad_covar}
\alias{reciprocity_min_dyad_covar}
\title{reciprocity_min_dyad_covar}
\usage{
reciprocity_min_dyad_covar(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Do individuals move to destinations that send more individuals to ego’s origin?
This is weighted by a dydic covariate.
This version of the effect is the minimum of the moves in either direction,
thereby guarding against degeneracy and guaranteeing sample size consistency.
It counts the ‘raw’ number of reciprocated transitions in the mobility network.
}
\keyword{internal}
