% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_effectFunctions_reciprocity.R
\name{reciprocity_min_resource_covar}
\alias{reciprocity_min_resource_covar}
\title{reciprocity_min_resource_covar}
\usage{
reciprocity_min_resource_covar(
  dep.var = 1,
  resource.attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F
)
}
\arguments{
\item{dep.var}{}

\item{resource.attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Do people reciprocate moves to other locations specifically if they and others
have a higher value on some covariate? For example, do women move to organisations
that send women to their origin organisation?
}
\keyword{internal}
