% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_effectFunctions_exogenous.R
\name{resource_covar_to_node_covar}
\alias{resource_covar_to_node_covar}
\title{resource_covar_to_node_covar}
\usage{
resource_covar_to_node_covar(
  dep.var = 1,
  resource.attribute.index,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  loop.contribution = F,
  getTargetContribution = F
)
}
\arguments{
\item{dep.var}{}

\item{resource.attribute.index}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{loop.contribution}{}

\item{getTargetContribution}{}
}
\value{
None.
}
\description{
resource_covar_to_node_covar
}
\keyword{internal}
