% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{simulateMobilityNetworks}
\alias{simulateMobilityNetworks}
\alias{simulateDistributionNetworks}
\title{simulateMobilityNetworks}
\usage{
simulateMobilityNetworks(
  cache,
  state,
  effects,
  ans,
  allowLoops,
  burnin,
  thinning,
  nSimulations,
  dep.var
)

simulateDistributionNetworks(
  cache,
  state,
  effects,
  ans,
  allowLoops,
  burnin,
  thinning,
  nSimulations,
  dep.var
)
}
\arguments{
\item{cache}{}

\item{state}{}

\item{effects}{}

\item{ans}{}

\item{allowLoops}{}

\item{burnin}{}

\item{thinning}{}

\item{nSimulations}{}

\item{dep.var}{}
}
\description{
simulateMobilityNetworks

simulateDistributionNetworks
}
