% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{staying_by_prop_bin_inflow}
\alias{staying_by_prop_bin_inflow}
\title{staying_by_prop_bin_inflow}
\usage{
staying_by_prop_bin_inflow(
  dep.var = 1,
  resource.attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{resource.attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Is the tendency to stay in vs. move out of a location dependent on the proportion
of individuals of type w that enter the location? This is especially geared
towards modelling how some locations become more or less
attractive dependent on the change in composition.
}
\keyword{internal}
