% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2x_effects_test.R
\name{target_change_match}
\alias{target_change_match}
\title{target_change_match}
\usage{
target_change_match(edge, j_new, effectName, ...)
}
\arguments{
\item{edge}{}

\item{j_new}{}

\item{effectName}{}

\item{...}{}
}
\value{
a vector containing (the change stat),(the target stat before change)
(the target stat after change),(whether this matches exactly),
(whether it matches rounded)
}
\description{
Function to test effects
}
\details{
Tests for specific resources and new proposed destination whether
the change statistic and the difference between two target statistics
before and after change match
}
\keyword{internal}
