% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2x_effects_test.R
\name{test_effect}
\alias{test_effect}
\title{test_effect

Test for each person in the example data and one randomly selected
alternative destination whether change and target statistics match.}
\usage{
test_effect(effectName, ...)
}
\arguments{
\item{effectName}{}

\item{...}{}
}
\value{
A data frame containing for each person in the example data:
Origin; Destination; Suggested destination; Change statistic;
Target before change; Target after change; Whether this is an exact match;
Whether this is a match rounded to 8 digits.
In case the last column contains values that are not all "1",
the tested function does not work properly.
}
\description{
test_effect

Test for each person in the example data and one randomly selected
alternative destination whether change and target statistics match.
}
\examples{
test_loop <- test_effect("loops")
}
\keyword{internal}
