% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{tie_weights_sigmoid}
\alias{tie_weights_sigmoid}
\title{tie_weights_sigmoid}
\usage{
tie_weights_sigmoid(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  alpha
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{alpha}{}
}
\description{
tie_weights_sigmoid
}
\keyword{internal}
