% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2e_effectFunctions_transitivity.R
\name{transitivity_AC}
\alias{transitivity_AC}
\title{transitivity_AC}
\usage{
transitivity_AC(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE,
  alpha = 1.1
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{alpha}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Is mobility clustered in groups? This is represented by the count of
transitive triads among three nodes, where the number of two-paths is
geometrically weighted down to avoid degeneracy.
}
\keyword{internal}
