% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2e_effectFunctions_transitivity.R
\name{transitivity_basic}
\alias{transitivity_basic}
\title{transitivity_basic}
\usage{
transitivity_basic(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Is mobility clustered in groups? This is represented by the count of
transitive triads among three nodes. This effect is prone to degeneracy.
}
\keyword{internal}
