% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_effectFunctions_exogenous.R
\name{alter_covariate}
\alias{alter_covariate}
\title{alter_covariate}
\usage{
alter_covariate(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  loop.contribution = T,
  getTargetContribution = F
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{loop.contribution}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Are locations higher on some attribute v more popular targets of mobility?
E.g., do workers have a tendency to move to larger organisations?
}
\keyword{internal}
