% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_auxiliaryFunctions.R
\name{autoCorrelationTest}
\alias{autoCorrelationTest}
\title{autoCorrelationTest}
\usage{
autoCorrelationTest(ans)
}
\arguments{
\item{ans}{An object of class "result.monan" resulting from an estimation with the function \code{\link[=estimateMobilityNetwork]{estimateMobilityNetwork()}}.}
}
\value{
A number indicating the auto-correlation.
}
\description{
The autoCorrelationTest indicates the degree to which the values of the dependent
variable of consecutive draws from the chain in phase 3 are correlated. Here lower
values are better. Values above 0.5 are very problematic and indicate that a
higher thinning is needed.
}
\examples{
# regression diagnostics
autoCorrelationTest(myResDN)
}
