% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{autoCorrelationTest}
\alias{autoCorrelationTest}
\title{Title}
\usage{
autoCorrelationTest(dep.var, ans)
}
\arguments{
\item{ans}{}
}
\description{
Title
}
\examples{
# regression diagnostics
autoCorrelationTest(exampleDependentVariable, exampleResDN)
}
\seealso{
\code{\link[=estimateMobilityNetwork]{estimateMobilityNetwork()}}
}
