% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c_effectFunctions_concentration.R
\name{concentration_GW}
\alias{concentration_GW}
\title{concentration_GW}
\usage{
concentration_GW(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  lambda = 2
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{lambda}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Is there a bandwagon effect in mobility, i.e. do mobile individuals move to locations
that are the destination of many others from their origin? The functional form of this
statistic assumes that there are decreasing additional returns to more others on the
same mobility path. For example, the probability to choose a mobility path that already
contains 20 other individuals is hardly different from a path with 25 other individuals;
however, there is a substantial difference in the comparison of paths with 2 or 7 other
individuals.
}
\keyword{internal}
