% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c_effectFunctions_concentration.R
\name{concentration_GW_dyad_covar_bin}
\alias{concentration_GW_dyad_covar_bin}
\title{concentration_GW_dyad_covar_bin}
\usage{
concentration_GW_dyad_covar_bin(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  lambda = 2
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{lambda}{}
}
\description{
concentration_GW_dyad_covar_bin
}
\keyword{internal}
