% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{createEdgelist}
\alias{createEdgelist}
\title{Create an edgelist object}
\usage{
createEdgelist(el, nodeSet = c("actors", "actors", "edges"))
}
\arguments{
\item{el}{an edgelist in the form of a matrix with two columns and N rows.
The first column indicates the origin of a person, the second row the destination.
Each row is one observation.}

\item{nodeSet}{}
}
\value{
an object of type edgelist.monan
}
\description{
Create an edgelist object
}
\examples{
# create an object of class edgelist.monan
transfers <- createEdgelist(mobilityEdgelist, nodeSet = c("locations", "locations", "people"))
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}
}
