% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{createEdgelist}
\alias{createEdgelist}
\title{createEdgelist}
\usage{
createEdgelist(el, nodeSet = c("location", "location", "individuals"))
}
\arguments{
\item{el}{An edgelist in the form of a matrix with two columns and N rows.
The first column indicates the origin of a person/resource, the second row the destination.
Each row is one observation.}

\item{nodeSet}{The nodesets of the edgelists. For edgelists, this is a vector with three
entries describing between what locations who is mobile, i.e., c(location, location, individuals/resources).}
}
\value{
an object of type edgelist.monan
}
\description{
Creates an edgelist object, which is the standard format of the outcome to be modelled
by monan.
}
\examples{
# create an object of class edgelist.monan
transfers <- createEdgelist(mobilityEdgelist, c("organisations", "organisations", "people"))
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}
}
