% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{createEdgelist}
\alias{createEdgelist}
\title{Create an edgelist object}
\usage{
createEdgelist(el, nodeSet = c("actors", "actors", "edges"))
}
\arguments{
\item{nodeSet}{}
}
\description{
Create an edgelist object
}
\examples{
# create an object of class edgelist.monan
transfers_EL <- createEdgelist(mobilityEdgelist, nodeSet = c("location", "location", "people"))
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}
}
