% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{createEffectsObject}
\alias{createEffectsObject}
\title{Title}
\usage{
createEffectsObject(effectInit, checkProcessState = NULL)
}
\arguments{
\item{checkProcessState}{}
}
\description{
Title
}
\examples{
# create an effects object
myEffects <- createEffectsObject(
  list(
    list("loops"),
    list("min_reciprocity"),
    list("dyadic_covariate", attribute.index = "sameRegion"),
    list("alter_covariate",  attribute.index = "size"),
    list("resource_covar_to_node_covar", attribute.index = "region", resource.attribute.index = "sex"),
    list("loops_resource_covar", resource.attribute.index = "sex")
  )
)
}
