% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{createEffectsObject}
\alias{createEffectsObject}
\title{createEffectsObject}
\usage{
createEffectsObject(effectInit, checkProcessState = NULL)
}
\arguments{
\item{effectInit}{A list of "Effects", where each effect to be included is specified as a list
that contains the effect name and the additional parameters it needs.}

\item{checkProcessState}{For internal use only.}
}
\description{
Specifies the model with endogenous and exogenous predictors.
The predictors in the model are called “Effects”.
}
\examples{
# create an effects object
myEffects <- createEffectsObject(
  list(
    list("loops"),
    list("min_reciprocity"),
    list("dyadic_covariate", attribute.index = "sameRegion"),
    list("alter_covariate",  attribute.index = "size"),
    list("resource_covar_to_node_covar", attribute.index = "region", 
          resource.attribute.index = "sex"),
    list("loops_resource_covar", resource.attribute.index = "sex")
  )
)
}
