% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{createNetwork}
\alias{createNetwork}
\alias{dyadicCovar}
\title{createNetwork}
\usage{
createNetwork(
  m,
  isSymmetric = FALSE,
  isBipartite = FALSE,
  nodeSet = NULL,
  nodes = NULL
)

dyadicCovar(
  m,
  isSymmetric = FALSE,
  isBipartite = FALSE,
  nodeSet = NULL,
  nodes = NULL
)
}
\arguments{
\item{m}{A square matrix containing the network data.}

\item{isSymmetric}{Currently not in use.}

\item{isBipartite}{Currently not in use.}

\item{nodeSet}{Which nodeset are the nodes of the network. Usually this will
be the locations in the data.}

\item{nodes}{Alternative way to specify the nodeSet by naming nodes:
nodes denote the locations in the edgelist}
}
\value{
An object of class "network.monan".
}
\description{
Defines a network between locations, generally to be used as a predictor in the model.
NOTE: The outcome variable of the model is not defined as a network, but as an edgelist!
}
\examples{
# create an object of class network.monan
sameRegion <- outer(orgRegion, orgRegion, "==") * 1
sameRegion <- createNetwork(sameRegion, nodeSet = c("organisations", "organisations"))
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}, \code{\link[=createEdgelist]{createEdgelist()}}
}
