% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{createNetwork}
\alias{createNetwork}
\title{Title}
\usage{
createNetwork(
  m,
  isSymmetric = FALSE,
  isBipartite = FALSE,
  nodeSet = c("actors", "actors")
)
}
\arguments{
\item{nodeSet}{}
}
\description{
Title
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}
}
