% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{createNodeSet}
\alias{createNodeSet}
\title{Title}
\usage{
createNodeSet(x = NULL, isPresent = NULL, considerWhenSampling = NULL)
}
\arguments{
\item{considerWhenSampling}{}
}
\description{
Title
}
\examples{
# create an object of class nodeSet.monan
people_NS <- createNodeSet(1:nrow(mobilityEdgelist))
location_NS <- createNodeSet(1:length(nodeVarCat))
}
\seealso{
\code{\link{createProcessState}}
}
