% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{createNodeVariable}
\alias{createNodeVariable}
\title{Title}
\usage{
createNodeVariable(
  values,
  range = NULL,
  nodeSet = "actors",
  addSame = FALSE,
  addSim = FALSE
)
}
\arguments{
\item{addSim}{}
}
\description{
Title
}
\examples{
# create an object of class nodeVar.monan
region <- createNodeVariable(orgRegion, nodeSet = "locations")
size <- createNodeVariable(orgSize, nodeSet = "locations", addSim = TRUE)
sex <- createNodeVariable(indSex, nodeSet = "people")
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}
}
