% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{createNodeVariable}
\alias{createNodeVariable}
\title{Title}
\usage{
createNodeVariable(
  values,
  range = NULL,
  nodeSet = "actors",
  addSame = F,
  addSim = F
)
}
\arguments{
\item{addSim}{}
}
\description{
Title
}
