% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{createProcessState}
\alias{createProcessState}
\title{Title}
\usage{
createProcessState(elements)
}
\arguments{
\item{elements}{}
}
\description{
Title
}
\examples{
# Create a process state out of the mobility data objects:
# create objects (which are later combined to the process state)
transfers <- createEdgelist(mobilityEdgelist, nodeSet = c("locations", "locations", "people"))
people <- createNodeSet(1:nrow(mobilityEdgelist))
locations <- createNodeSet(1:length(nodeVarCat))
sameRegion <- outer(nodeVarCat, nodeVarCat, "==")*1
sameRegion <- createNetwork(sameRegion, nodeSet = c("locations", "locations"))
region <- createNodeVariable(nodeVarCat, nodeSet = "locations")
size <- createNodeVariable(nodeVarCont, nodeSet = "locations", addSim = TRUE)
sex <- createNodeVariable(resVarCat, nodeSet = "people")

# combine created objects to the process state
myState <- createProcessState(list(
  transfers = transfers,
  
  people = people,
  locations = locations,
  
  sameRegion = sameRegion,
  region = region,
  size = size,
  sex = sex
))
}
\seealso{
\code{\link[=createEdgelist]{createEdgelist()}}, \code{\link[=createNodeSet]{createNodeSet()}},
\code{\link[=createNodeVariable]{createNodeVariable()}}, \code{\link[=createNetwork]{createNetwork()}}
}
