% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{createProcessState}
\alias{createProcessState}
\title{Title}
\usage{
createProcessState(elements)
}
\arguments{
\item{elements}{}
}
\description{
Title
}
\examples{
# Create a process state out of the example data objects:
# create objects (which are later combined to the process state)
transfers_EL <- createEdgelist(mobilityEdgelist, nodeSet = c("location", "location", "people"))
people_NS <- createNodeSet(1:nrow(mobilityEdgelist))
location_NS <- createNodeSet(1:length(nodeVarCat))
nodeVarCat_NV <- createNodeVariable(nodeVarCat, nodeSet = "location")
nodeVarCont_NV <- createNodeVariable(nodeVarCont, nodeSet = "location", addSim = TRUE)
resVarCat_NV <- createNodeVariable(resVarCat, nodeSet = "people")

# combine created objects to the process state
exampleState <- createProcessState(list(
  
  transfers = transfers_EL,
  people = people_NS,
  location = location_NS,
  nodeVarCat = nodeVarCat_NV,
  nodeVarCont = nodeVarCont_NV,
  resVarCat = resVarCat_NV
  
))
}
\seealso{
\code{\link[=createEdgelist]{createEdgelist()}}, \code{\link[=createNodeSet]{createNodeSet()}},
\code{\link[=createNodeVariable]{createNodeVariable()}}, \code{\link[=createNetwork]{createNetwork()}}
}
