% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{createWeightedCache}
\alias{createWeightedCache}
\title{Title}
\usage{
createWeightedCache(processState, cacheObjectNames, resourceCovariates = NULL)
}
\arguments{
\item{resourceCovariates}{}
}
\description{
Title
}
\examples{
# define dependent variable and create cache object
myDependentVariable <- "transfers"
myCache <- createWeightedCache(myState, myDependentVariable, resourceCovariates = c("sex"))
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}
}
