% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{createWeightedCache}
\alias{createWeightedCache}
\title{Title}
\usage{
createWeightedCache(processState, cacheObjectNames, resourceCovariates = NULL)
}
\arguments{
\item{processState}{}

\item{cacheObjectNames}{}

\item{resourceCovariates}{}
}
\description{
Title
}
\examples{
# define dependent variable and create cache object
exampleDependentVariable <- "transfers"

exampleCache <- createWeightedCache(exampleState, exampleDependentVariable, resourceCovariates = c("resVarCat"))
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}
}
