% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{createWeightedCache}
\alias{createWeightedCache}
\title{Title}
\usage{
createWeightedCache(processState, cacheObjectNames, ressourceCovariates = NULL)
}
\arguments{
\item{processState}{}

\item{cacheObjectNames}{}

\item{ressourceCovariates}{}
}
\description{
Title
}
