% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R
\name{createWeightedCache}
\alias{createWeightedCache}
\title{createWeightedCache}
\usage{
createWeightedCache(processState, resourceCovariates = NULL)
}
\arguments{
\item{processState}{Outdated.}

\item{resourceCovariates}{Outdated.}
}
\value{
Outdated.
}
\description{
Since MoNAn version 1.0.0, this function no longer exists.
}
