% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{crowding_out_prop_covar_bin}
\alias{crowding_out_prop_covar_bin}
\title{crowding_out_prop_covar_bin}
\usage{
crowding_out_prop_covar_bin(
  dep.var = 1,
  resource.attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F
)
}
\arguments{
\item{dep.var}{}

\item{resource.attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Is the tendency to stay in vs. move out of a location of individuals of type
non-w dependent on the proportion of individuals of type w moving into the location?
This is especially geared towards modelling how some locations become more or
less attractive dependent on the change in composition for
particular groups. This models segregation dynamics.
}
\keyword{internal}
