% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_effectFunctions_exogenous.R
\name{dyadic_covariate}
\alias{dyadic_covariate}
\title{dyadic_covariate}
\usage{
dyadic_covariate(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Is mobility between locations predicted by the dyadic covariate u? E.g., is mobility
likely between organisations that are in the same region? Note that in many cases
dyadic covariates can convey the same information as
the ‘same covariate’ or the ‘covariate similarity’ effects.
}
\keyword{internal}
