% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_effectFunctions_exogenous.R
\name{dyadic_covariate_resource_attribute}
\alias{dyadic_covariate_resource_attribute}
\title{dyadic_covariate_resource_attribute}
\usage{
dyadic_covariate_resource_attribute(
  dep.var = 1,
  attribute.index,
  resource.attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{resource.attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
None.
}
\description{
dyadic_covariate_resource_attribute
}
\keyword{internal}
