% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_coreFunctions.R, R/4_auxiliaryFunctions.R
\name{estimateMobilityNetwork}
\alias{estimateMobilityNetwork}
\alias{estimateDistributionNetwork}
\alias{print.result.monan}
\title{estimateMobilityNetwork}
\usage{
estimateMobilityNetwork(
  dep.var,
  state,
  cache,
  effects,
  algorithm,
  initialParameters = NULL,
  prevAns = NULL,
  parallel = FALSE,
  cpus = 3,
  verbose = FALSE,
  returnDeps = FALSE,
  multinomialProposal = FALSE,
  fish = FALSE
)

estimateDistributionNetwork(
  dep.var,
  state,
  cache,
  effects,
  algorithm,
  initialParameters = NULL,
  prevAns = NULL,
  parallel = FALSE,
  cpus = 3,
  verbose = FALSE,
  returnDeps = FALSE,
  multinomialProposal = FALSE,
  fish = FALSE
)

\method{print}{result.monan}(x, covMat = FALSE, ...)
}
\arguments{
\item{dep.var}{The outcome variable that is modelled.}

\item{state}{A monan state object that contains all relevant information about
the outcome in the form of an edgelist, the nodesets, and covariates.}

\item{cache}{A monan cache object created from the same state object that is
used in the estimation}

\item{effects}{An effect object that specifies the model.}

\item{initialParameters}{Starting values for the parameters. Using starting
values e.g. from a multinomial logit model (see getMultinomialStatistics())
increases the chances of model convergence in the first run of the estimation
considerably.}

\item{prevAns}{If a previous estimation did not yield satisfactory convergence,
the outcome object of that estimation should be specified here to provide new
starting values for the estimation.}

\item{parallel}{Computation on multiple cores?}

\item{cpus}{Number of cores for computation in case parallel = T.}

\item{verbose}{Logical: display information about estimation progress in the console?}

\item{returnDeps}{Logical: should the simulated values of Phase 3 be stored and returned?
This is necessary to run GoF tests.
Note that this might result in very large objects.}

\item{multinomialProposal}{How should the next possible outcome in the simulation chains
be sampled? If TRUE, fewer simulation steps are needed, but each simulation
step takes considerably longer.}

\item{fish}{Logical: display a fish?}

\item{x}{An ans object resulting from an estimation with estimateMobilityNetwork()}

\item{covMat}{Logical indicating whether the covariance matrix should be printed}

\item{...}{For internal use only}
}
\value{
A monan results object that contains the estimates, standard errors,
and convergence statistics. Furthermore, the covariance matrix used to calculate
the standard errors is included, which also shows colinearity between effects.
In case returnDeps = TRUE, the simulations of Phase 3 are included, too.
}
\description{
The core function of the package in which the model for the analysis of
mobility tables is estimated.
}
\examples{
\dontrun{
# estimate mobility network
myResDN <- estimateMobilityNetwork(myDependentVariable,
  myState, myCache, myEffects, myAlg,
  initialParameters = NULL,
  parallel = TRUE, cpus = 4,
  verbose = TRUE,
  returnDeps = TRUE,
  multinomialProposal = FALSE,
  fish = FALSE
)

myResDN

# estimate mobility network again based on previous results to improve convergence
myResDN <- estimateMobilityNetwork(myDependentVariable,
  myState, myCache, myEffects, myAlg,
  prevAns = myResDN,
  parallel = TRUE, cpus = 4,
  verbose = TRUE,
  returnDeps = TRUE,
  multinomialProposal = FALSE,
  fish = FALSE
)
}
myResDN
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}, \code{\link[=createWeightedCache]{createWeightedCache()}},
\code{\link[=createEffectsObject]{createEffectsObject()}}, \link{createAlgorithm}
}
