% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{estimateMobilityNetwork}
\alias{estimateMobilityNetwork}
\alias{estimateDistributionNetwork}
\title{Title}
\usage{
estimateMobilityNetwork(
  dep.var,
  state,
  cache,
  effects,
  initialParameters = NULL,
  prevAns = NULL,
  burnInN1 = 500,
  iterationsN1 = 200,
  thinningN1 = 20,
  gainN1 = 1,
  burnInN2 = burnInN1,
  nsubN2 = 6,
  initGain = 0.02,
  thinningN2 = 100,
  initialIterationsN2 = 50,
  iterationsN3 = 1000,
  burnInN3 = 2000,
  thinningN3 = 30,
  allowLoops = NULL,
  parallel = F,
  cpus = 3,
  verbose = F,
  returnDeps = F,
  multinomialProposal = F,
  fish = F
)
}
\arguments{
\item{fish}{}
}
\description{
Title
}
\examples{
# estimate mobility network
exampleResDN <- estimateMobilityNetwork(dependentVariable, 
exampleState, exampleCache, exampleEffects,
initialParameters = NULL,
burnInN1 = 200, iterationsN1 = 50, thinningN1 = 2000, gainN1 = 0.1,
burnInN2 = 2000, nsubN2 = 4, initGain = 0.25, thinningN2 = 2000, 
initialIterationsN2 = 25,
iterationsN3 = 500, burnInN3 = 2000, thinningN3 = 2000,
parallel = T, cpus = 4,
allowLoops = T,
verbose = T,
returnDeps = T,
multinomialProposal = T,
fish = F)
exampleResDN
}
\seealso{
\code{\link[=createProcessState]{createProcessState()}}, \code{\link[=createWeightedCache]{createWeightedCache()}},
\code{\link[=createEffectsObject]{createEffectsObject()}}
}
