% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{extractTraces}
\alias{extractTraces}
\alias{plot.traces.monan}
\title{Title}
\usage{
extractTraces(dep.var, ans, effects)

\method{plot}{traces.monan}(xx, ...)
}
\arguments{
\item{effects}{}

\item{...}{}
}
\description{
Title

Title
}
\examples{
# regression diagnostics
traces <- extractTraces(dependentVariable, resDN, exampleEffects)
plot(traces)
}
\seealso{
\code{\link{estimateMobilityNetwork}}, \code{\link{createEffectsObject}}
}
