% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_auxiliaryFunctions.R
\name{extractTraces}
\alias{extractTraces}
\alias{plot.traces.monan}
\title{Title}
\usage{
extractTraces(dep.var, ans, effects)

\method{plot}{traces.monan}(xx, ...)
}
\arguments{
\item{effects}{}

\item{...}{}
}
\description{
Title

Title
}
\examples{
# regression diagnostics
exampleTraces <- extractTraces(exampleDependentVariable, exampleResDN, exampleEffects)
plot(exampleTraces)
}
\seealso{
\code{\link[=estimateMobilityNetwork]{estimateMobilityNetwork()}}, \code{\link[=createEffectsObject]{createEffectsObject()}}
}
