% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_auxiliaryFunctions.R
\name{gofDistributionNetwork}
\alias{gofDistributionNetwork}
\alias{plot.gof.stats.monan}
\title{gofDistributionNetwork}
\usage{
gofDistributionNetwork(
  ans,
  simulations,
  gofFunction,
  dep.var = NULL,
  lvls = NULL
)

\method{plot}{gof.stats.monan}(x, lvls, ...)
}
\arguments{
\item{ans}{The ans object resulting from an estimation with "estimateMobilityNetwork"}

\item{simulations}{The simulated outcomes with which the observed statistics are compared.
Usually, they are stored in the ans$deps, in case deps = T was specified in the
estimation.}

\item{gofFunction}{A gof function that specifies which auxiliary outcome should be,
e.g., "getIndegree" or "getTieWeights"}

\item{dep.var}{The dependent variable specified in the estimation}

\item{lvls}{The values for which the gofFunction should be calculated / plotted}

\item{x}{A gof.stats.monan object, usually obtained from running "gofDistributionNetwork"}

\item{...}{Additional plotting parameters, use discouraged.}
}
\value{
A list containing (1) the observed values of the auxiliary statistics and
(2) a list of the simulated values of the auxiliary statistics
}
\description{
Akin to ERGMs, goodness of fit testing is available to see whether auxiliary
statistics are well captured by the model. The logic behind gof testing for network models is outlined in
Hunter et al. (2008) and Lospinoso and Snijders (2019).
}
\examples{
# goodness of fit
myGofIndegree <- gofDistributionNetwork(ans = myResDN, 
                                        simulations = myResDN$deps, 
                                        gofFunction = getIndegree, 
                                        lvls = 1:100)

myGofTieWeight <- gofDistributionNetwork(ans = myResDN, 
                                         simulations = myResDN$deps, 
                                         gofFunction = getTieWeights, 
                                         lvls = 1:30)
plot(myGofIndegree,  lvls = 20:70)
plot(myGofTieWeight, lvls = 1:15)
}
\references{
Hunter, D. R., Goodreau, S. M., & Handcock, M. S. (2008). Goodness of fit of social network models.
Journal of the american statistical association, 103(481), 248-258.

Lospinoso, J., & Snijders, T. A. (2019).
Goodness of fit for stochastic actor-oriented models. Methodological Innovations, 12(3).
}
\seealso{
\code{\link[=estimateMobilityNetwork]{estimateMobilityNetwork()}}, \code{\link[=getIndegree]{getIndegree()}},
\code{\link[=getTieWeights]{getTieWeights()}}
}
