% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_auxiliaryFunctions.R
\name{gofDistributionNetwork}
\alias{gofDistributionNetwork}
\alias{plot.gof.stats.monan}
\title{Title}
\usage{
gofDistributionNetwork(
  ans,
  simulations,
  gofFunction,
  dep.var = NULL,
  lvls = NULL
)

\method{plot}{gof.stats.monan}(x, ...)
}
\arguments{
\item{lvls}{}

\item{x}{a gofObject}
}
\description{
Title

Title
}
\examples{
# goodness of fit
myGofIndegree <- gofDistributionNetwork(ans = myResDN, simulations = myResDN$deps, gofFunction = getIndegree, lvls = 1:100)
plot(myGofIndegree)

myGofTieWeight <- gofDistributionNetwork(ans = myResDN, simulations = myResDN$deps, gofFunction = getTieWeights, lvls = 1:30)
plot(myGofTieWeight)
}
\seealso{
\code{\link[=estimateMobilityNetwork]{estimateMobilityNetwork()}}, \code{\link[=getIndegree]{getIndegree()}},
\code{\link[=getTieWeights]{getTieWeights()}}
}
