% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{gofDistributionNetwork}
\alias{gofDistributionNetwork}
\alias{plot.gof.stats.monan}
\title{Title}
\usage{
gofDistributionNetwork(
  ans,
  simulations,
  gofFunction,
  dep.var = NULL,
  lvls = NULL
)

\method{plot}{gof.stats.monan}(gofObject, lvls = NULL)
}
\arguments{
\item{lvls}{}
}
\description{
Title

Title
}
\examples{
# goodness of fit
exampleGofIndegree <- gofDistributionNetwork(ans = exampleResDN, simulations = exampleResDN$deps, gofFunction = getIndegree, lvls = 1:100)
plot(exampleGofIndegree)

exampleGofTieWeight <- gofDistributionNetwork(ans = exampleResDN, simulations = exampleResDN$deps, gofFunction = getTieWeights, lvls = 1:30)
plot(exampleGofTieWeight)
}
\seealso{
\code{\link[=estimateMobilityNetwork]{estimateMobilityNetwork()}}, \code{\link[=getIndegree]{getIndegree()}},
\code{\link[=getTieWeights]{getTieWeights()}}
}
