% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{gofDistributionNetwork}
\alias{gofDistributionNetwork}
\title{Title}
\usage{
gofDistributionNetwork(
  ans,
  simulations,
  gofFunction,
  dep.var = NULL,
  lvls = NULL
)
}
\arguments{
\item{lvls}{}
}
\description{
Title
}
\examples{
# goodness of fit
gof.indegree.1 <- gofDistributionNetwork(ans = resDN, simulations = resDN$deps, gofFunction = getIndegree, lvls = 1:100)
plot(gof.indegree.1)

gof.tie.weights.1 <- gofDistributionNetwork(ans = resDN, simulations = resDN$deps, gofFunction = getTieWeights, lvls = 1:30)
plot(gof.tie.weights.1)
}
\seealso{
\code{\link{estimateMobilityNetwork}}, \code{\link{getIndegree}}, 
\code{\link{getTieWeights}}
}
