% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{loops}
\alias{loops}
\title{loops}
\usage{
loops(dep.var = 1, state, cache, i, j, edge, update, getTargetContribution = F)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\description{
loops
}
\keyword{internal}
