% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_effectFunctions_loops.R
\name{loops}
\alias{loops}
\title{loops}
\usage{
loops(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Do individuals stay in their location of origin, compared to going to a different location?
}
\keyword{internal}
