% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_effectFunctions.R
\name{loops_GW}
\alias{loops_GW}
\title{loops_GW}
\usage{
loops_GW(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F,
  alpha = 2
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}

\item{alpha}{}
}
\description{
loops_GW
}
\keyword{internal}
