% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_effectFunctions_loops.R
\name{loops_node_covar}
\alias{loops_node_covar}
\title{loops_node_covar}
\usage{
loops_node_covar(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Are locations with specific attributes ‘stickier’ than others, i.e., do individuals
have a higher propensity to stay in some locations? E.g., are individuals working
in organisations in one region less likely to change their employer?
}
\keyword{internal}
