% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_effectFunctions_loops.R
\name{loops_resource_covar}
\alias{loops_resource_covar}
\title{loops_resource_covar}
\usage{
loops_resource_covar(
  dep.var = 1,
  resource.attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{resource.attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Are individuals with certain characteristics more likely to remain in their current
location? For example, are men more likely to remain in their current organisation,
while women are more likely to move employer?
}
\keyword{internal}
