% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilityData.R
\docType{data}
\name{mobilityData}
\alias{mobilityData}
\alias{mobilityEdgelist}
\alias{orgRegion}
\alias{orgSize}
\alias{indSex}
\title{Example Data for the MoNAn Package}
\format{
\code{mobilityEdgelist}
A data frame with 743 rows and 2 columns.

\code{orgRegion}
An object with 17 values.

\code{orgSize}
An object with 17 values.

\code{indSex}
An object with 742 values.
}
\usage{
mobilityEdgelist

orgRegion

orgSize

indSex
}
\description{
These are example data for the MoNAn package and can be used to estimate a
mobility network. The following objects are provided for this purpose:

\describe{
\item{mobilityEdgelist}{blabla}
\item{orgRegion}{blabla}
\item{orgSize}{blabla}
\item{indSex}{blabla}
}
}
\keyword{datasets}
