% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoNAn-package.R
\docType{package}
\name{MoNAn-package}
\alias{MoNAn}
\alias{MoNAn-package}
\title{MoNAn: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Per Block \email{block@soziologie.uzh.ch}

}
\keyword{internal}
