% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myOutcomeObjects.R
\docType{data}
\name{myOutcomeObjects}
\alias{myOutcomeObjects}
\alias{myState}
\alias{myDependentVariable}
\alias{myCache}
\alias{myEffects}
\alias{myAlg}
\alias{myResDN}
\title{Exemplary Outcome Objects for the MoNAn Package}
\format{
\code{myState}
An object of class "processState.monan".

\code{myDependentVariable}
An object of class "character".

\code{myCache}
An object of class "list".

\code{myEffects}
An object of class "effectsList.monan".

\code{myAlg}
An object of class "algorithm.monan".

\code{myResDN}
An object of class "result.monan".
}
\usage{
myState

myDependentVariable

myCache

myEffects

myAlg

myResDN
}
\description{
These are exemplary outcome objects for the MoNAn package and can be used in
order not to run all precedent functions and thus save time. The following
products are provided:

\describe{
\item{myState}{blabla}
\item{myDependentVariable}{blabla}
\item{myCache}{blabla}
\item{myEffects}{blabla}
\item{myAlg}{blabla}
\item{myResDN}{blabla}
}
}
\keyword{datasets}
