% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myOutcomeObjects.R
\docType{data}
\name{myOutcomeObjects}
\alias{myOutcomeObjects}
\alias{myState}
\alias{myCache}
\alias{myEffects}
\alias{myAlg}
\alias{myResDN}
\alias{mySimDN}
\title{Exemplary Outcome Objects for the MoNAn Package}
\format{
\code{myState}
An object of class \code{processState.monan} created by the function \code{\link[=createProcessState]{createProcessState()}}.

\code{myCache}
An object of class \code{list} created by the function \code{\link[=createWeightedCache]{createWeightedCache()}}.

\code{myEffects}
An object of class \code{effectsList.monan} created by the function \code{\link[=createEffectsObject]{createEffectsObject()}}.

\code{myAlg}
An object of class \code{algorithm.monan} created by the function \code{\link[=createAlgorithm]{createAlgorithm()}}.

\code{myResDN}
An object of class \code{result.monan} created by the function \code{\link[=estimateMobilityNetwork]{estimateMobilityNetwork()}}.

\code{mySimDN}
An object of class \code{sims.monan} created by the function \code{\link[=simulateMobilityNetworks]{simulateMobilityNetworks()}}.
}
\usage{
myState

myCache

myEffects

myAlg

myResDN

mySimDN
}
\description{
These are exemplary outcome objects for the MoNAn package and can be used in
order not to run all precedent functions and thus save time. The following
products are provided:
}
\keyword{datasets}
