% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_auxiliaryFunctions.R
\name{scoreTest}
\alias{scoreTest}
\alias{print.scoretest.monan}
\title{Title}
\usage{
scoreTest(dep.var, ans, effects)

\method{print}{scoretest.monan}(x, ...)
}
\arguments{
\item{effects}{}

\item{...}{}
}
\description{
Title

Title
}
\examples{
# test whether other effects should be included
exampleEffects2 <- createEffectsObject(
  list(
    list("loops"),
    list("min_reciprocity"),
    list("loops_resource_covar", resource.attribute.index = "resVarCat"),
    list("min_transitivity")
  )
)
exampleTestEffects2 <- scoreTest(exampleDependentVariable, exampleResDN, exampleEffects2)
exampleTestEffects2
}
\seealso{
\code{\link[=createEffectsObject]{createEffectsObject()}}, \code{\link[=estimateMobilityNetwork]{estimateMobilityNetwork()}}
}
