% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_auxiliaryFunctions.R
\name{scoreTest}
\alias{scoreTest}
\alias{print.scoretest.monan}
\title{Title}
\usage{
scoreTest(dep.var, ans, effects)

\method{print}{scoretest.monan}(x, ...)
}
\arguments{
\item{effects}{}

\item{...}{}
}
\description{
Title

Title
}
\examples{
# test whether other effects should be included
myEffects2 <- createEffectsObject(
  list(
    list("loops"),
    list("min_reciprocity"),
    list("dyadic_covariate", attribute.index = "sameRegion"),
    list("alter_covariate", attribute.index = "size"),
    list("resource_covar_to_node_covar", attribute.index = "region", resource.attribute.index = "sex"),
    list("loops_resource_covar", resource.attribute.index = "sex"),
    list("min_transitivity")
  )
)

test_ME.2 <- scoreTest(myDependentVariable, myResDN, myEffects2)
test_ME.2
}
\seealso{
\code{\link[=createEffectsObject]{createEffectsObject()}}, \code{\link[=estimateMobilityNetwork]{estimateMobilityNetwork()}}
}
