% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_effectFunctions_exogenous.R
\name{sim_covariate}
\alias{sim_covariate}
\title{sim_covariate}
\usage{
sim_covariate(
  dep.var = 1,
  attribute.index,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = F
)
}
\arguments{
\item{dep.var}{}

\item{attribute.index}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Is mobility more likely between locations that are similar on some attribute v?
E.g., is mobility more likely between organisations that have a similar size?
}
\keyword{internal}
