% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{simulateMobilityNetworks}
\alias{simulateMobilityNetworks}
\alias{simulateDistributionNetworks}
\title{Title}
\usage{
simulateMobilityNetworks(
  cache,
  state,
  effects,
  ans,
  allowLoops,
  burnin,
  thinning,
  nSimulations,
  dep.var
)
}
\arguments{
\item{dep.var}{}
}
\description{
Title
}
