% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2e_effectFunctions_transitivity.R
\name{transitivity_min}
\alias{transitivity_min}
\title{transitivity_min}
\usage{
transitivity_min(
  dep.var = 1,
  state,
  cache,
  i,
  j,
  edge,
  update,
  getTargetContribution = FALSE
)
}
\arguments{
\item{dep.var}{}

\item{state}{}

\item{cache}{}

\item{i}{}

\item{j}{}

\item{edge}{}

\item{update}{}

\item{getTargetContribution}{}
}
\value{
Returns the change statistic or target statistic of the effect for
internal use by the estimation algorithm.
}
\description{
Is mobility clustered in groups? This is represented by the minimum of reciprocated
mobility being present among three nodes. Using the minimum ensures that the effect
is not degenerate and it is sample size consistent.
}
\keyword{internal}
